docs: move list.js script into its own file
